scraper/requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
numpy>=1.24.0
imagehash>=4.3.0
